refactor: introduce utils for probs computation
